refactor(v1): simplify validate persistence layout
